+++ buildSrc/src/main/kotlin/com/danrusu/pods4k/immutableArrays/immutableArraysToStandardCollections/TransformationsToSetFileGenerator.kt
@@ -16,6 +16,7 @@ internal object TransformationsToSetFileGenerator {
         val fileSpec = createFile(ImmutableArrayConfig.packageName, "TransformationsToSet") {
             addToSet()
             addToMutableSet()
+            addToHashSet()
         }
         fileSpec.writeTo(File(destinationPath, ""))
     }
@@ -42,7 +43,11 @@ private fun FileSpec.Builder.addToSet() {
 private fun FileSpec.Builder.addToMutableSet() {
     for (baseType in BaseType.entries) {
         function(
-            kdoc = "See [Array.toMutableSet]",
+            kdoc = """
+                See [Array.toMutableSet]
+
+                If the iteration order doesn't matter then use [toHashSet] as that produces a more efficient mutable set.
+            """.trimIndent(),
             receiver = baseType.getGeneratedTypeName(),
             name = "toMutableSet",
             returns = ClassName("kotlin.collections", "MutableSet").parameterizedBy(baseType.type),
@@ -53,6 +58,22 @@ private fun FileSpec.Builder.addToMutableSet() {
             // Important: This needs to meet the same contract as what's promised by the standard library to maintain
             // iteration order since this library is documented as a replacement for read-only lists.
             statement("return asList().toMutableSet()")
+        }
+    }
+}
+
+private fun FileSpec.Builder.addToHashSet() {
+    for (baseType in BaseType.entries) {
+        function(
+            kdoc = "See [Array.toHashSet]",
+            receiver = baseType.getGeneratedTypeName(),
+            name = "toHashSet",
+            returns = ClassName("kotlin.collections", "HashSet").parameterizedBy(baseType.type),
+        ) {
+            if (baseType == BaseType.GENERIC) {
+                addTypeVariable(baseType.type as TypeVariableName)
+            }
+            statement("return asList().toHashSet()")
         }
     }
 }