+++ immutable-arrays/transformations-to-standard-collections/src/main/kotlin/com/danrusu/pods4k/immutableArrays/TransformationsToSet.kt
@@ -9,6 +9,7 @@ import kotlin.Float
 import kotlin.Int
 import kotlin.Long
 import kotlin.Short
+import kotlin.collections.HashSet
 import kotlin.collections.MutableSet
 import kotlin.collections.Set
 
@@ -59,45 +60,117 @@ public fun ImmutableDoubleArray.toSet(): Set<Double> = asList().toSet()
 
 /**
  * See [Array.toMutableSet]
+ *
+ * If the iteration order doesn't matter then use [toHashSet] as that produces a more efficient
+ * mutable set.
  */
 public fun <T> ImmutableArray<T>.toMutableSet(): MutableSet<T> = asList().toMutableSet()
 
 /**
  * See [Array.toMutableSet]
+ *
+ * If the iteration order doesn't matter then use [toHashSet] as that produces a more efficient
+ * mutable set.
  */
 public fun ImmutableBooleanArray.toMutableSet(): MutableSet<Boolean> = asList().toMutableSet()
 
 /**
  * See [Array.toMutableSet]
+ *
+ * If the iteration order doesn't matter then use [toHashSet] as that produces a more efficient
+ * mutable set.
  */
 public fun ImmutableByteArray.toMutableSet(): MutableSet<Byte> = asList().toMutableSet()
 
 /**
  * See [Array.toMutableSet]
+ *
+ * If the iteration order doesn't matter then use [toHashSet] as that produces a more efficient
+ * mutable set.
  */
 public fun ImmutableCharArray.toMutableSet(): MutableSet<Char> = asList().toMutableSet()
 
 /**
  * See [Array.toMutableSet]
+ *
+ * If the iteration order doesn't matter then use [toHashSet] as that produces a more efficient
+ * mutable set.
  */
 public fun ImmutableShortArray.toMutableSet(): MutableSet<Short> = asList().toMutableSet()
 
 /**
  * See [Array.toMutableSet]
+ *
+ * If the iteration order doesn't matter then use [toHashSet] as that produces a more efficient
+ * mutable set.
  */
 public fun ImmutableIntArray.toMutableSet(): MutableSet<Int> = asList().toMutableSet()
 
 /**
  * See [Array.toMutableSet]
+ *
+ * If the iteration order doesn't matter then use [toHashSet] as that produces a more efficient
+ * mutable set.
  */
 public fun ImmutableLongArray.toMutableSet(): MutableSet<Long> = asList().toMutableSet()
 
 /**
  * See [Array.toMutableSet]
+ *
+ * If the iteration order doesn't matter then use [toHashSet] as that produces a more efficient
+ * mutable set.
  */
 public fun ImmutableFloatArray.toMutableSet(): MutableSet<Float> = asList().toMutableSet()
 
 /**
  * See [Array.toMutableSet]
+ *
+ * If the iteration order doesn't matter then use [toHashSet] as that produces a more efficient
+ * mutable set.
  */
 public fun ImmutableDoubleArray.toMutableSet(): MutableSet<Double> = asList().toMutableSet()
+
+/**
+ * See [Array.toHashSet]
+ */
+public fun <T> ImmutableArray<T>.toHashSet(): HashSet<T> = asList().toHashSet()
+
+/**
+ * See [Array.toHashSet]
+ */
+public fun ImmutableBooleanArray.toHashSet(): HashSet<Boolean> = asList().toHashSet()
+
+/**
+ * See [Array.toHashSet]
+ */
+public fun ImmutableByteArray.toHashSet(): HashSet<Byte> = asList().toHashSet()
+
+/**
+ * See [Array.toHashSet]
+ */
+public fun ImmutableCharArray.toHashSet(): HashSet<Char> = asList().toHashSet()
+
+/**
+ * See [Array.toHashSet]
+ */
+public fun ImmutableShortArray.toHashSet(): HashSet<Short> = asList().toHashSet()
+
+/**
+ * See [Array.toHashSet]
+ */
+public fun ImmutableIntArray.toHashSet(): HashSet<Int> = asList().toHashSet()
+
+/**
+ * See [Array.toHashSet]
+ */
+public fun ImmutableLongArray.toHashSet(): HashSet<Long> = asList().toHashSet()
+
+/**
+ * See [Array.toHashSet]
+ */
+public fun ImmutableFloatArray.toHashSet(): HashSet<Float> = asList().toHashSet()
+
+/**
+ * See [Array.toHashSet]
+ */
+public fun ImmutableDoubleArray.toHashSet(): HashSet<Double> = asList().toHashSet()